chore: cleanup circular references
