+++ src/module/applications/index.ts
@@ -1,0 +1,4 @@
+export { CustomSkillMenuV2 as CustomSkillMenu } from './custom-skills-menu';
+export { CustomCurrencyMenuV2 as CustomCurrencyMenu } from './custom-currency-menu';
+export { TrackedCompendiumsMenu } from "./tracked-compendiums-menu.mjs";
+export { CompendiumBrowser } from "./compendium-browser.mjs";
