+++ src/module/constants/index.ts
@@ -17,3 +17,15 @@ export const SYSTEM_ID = 'cosmere-rpg';
  * Author name of the module.
  */
 export const AUTHOR_NAME = 'Khunkurisu';
+
+export const SETTINGS = {
+	CUSTOM_SKILLS: 'customSkills',
+	CUSTOM_CURRENCIES: 'customCurrency',
+	MENU_CUSTOM_CURRENCY: 'customCurrencyMenu',
+	MENU_CUSTOM_SKILL: 'customSkillMenu',
+	MENU_TRACKED_COMPENDIUMS: 'trackedCompendiumsMenu',
+	SHEET_ENCUMBRANCE_BAR_CLIENT: 'encumbranceBarClient',
+	SHEET_ENCUMBRANCE_BAR_WORLD: 'encumbranceBarWorld',
+	GENERAL_COMPENDIUM_MANAGER: 'compendiumManager',
+	GENERAL_TRACKED_COMPENDIUMS: 'trackedCompendiums',
+} as const;
