+++ src/module/sheets/actor-sheet-encumbrance-bar.mjs
@@ -1,5 +1,4 @@
-import { MODULE_ID } from "../constants";
-import { SETTINGS } from "../settings";
+import { MODULE_ID, SETTINGS } from "../constants";
 
 const templatePath = 'modules/cosmere-rpg-workbench/templates/sheets/parts/actor-encumbrance-bar.hbs';
